requirements.txt
selenium==3.141.0
requests==2.25.1
pytube3==9.6.4
Pillow==8.3.2
pytube==11.0.1
